Basic FurnaceFly Functionality
The furnaceFly prototype now exists in the MenuScene, with a few public fields that can be used to adjust its behavior where needed. The one thing I currently see remaining is the projectile rotation which will orient the projectile sprite according to its vector of motion which I will get done by tomorrow!